Add first structure and working functions in async and sync mode
