+++ pymarsys/__init__.py
@@ -1,2 +1,2 @@
 __version__ = 0.1
-from .sync.sync_emarsys import SyncEmarsys
+from .emarsys import Emarsys
