+++ requirements.txt
@@ -1,4 +1,4 @@
+aiohttp==1.0.5
+pytest==3.0.4
 requests==2.12.1
 responses==0.5.1
-pytest==3.0.4
-
